Refactor main.rs
