+++ src/handler.rs
@@ -1,0 +1,30 @@
+use serenity::model::application::interaction::{Interaction, InteractionType};
+use serenity::model::prelude::{Activity, Ready};
+use serenity::prelude::{Context, EventHandler};
+
+use crate::commands::BotCommandManager;
+use crate::error::DisplayErr;
+use crate::{DATABASE, STATUS_MESSAGE};
+
+/// An event handler for the bot.
+pub struct BotHandler;
+
+#[serenity::async_trait]
+impl EventHandler for BotHandler {
+    async fn ready(&self, ctx: Context, ready: Ready) {
+        ctx.set_activity(Activity::playing(STATUS_MESSAGE)).await;
+
+        let result = BotCommandManager::register_all(&ctx).await;
+        result.eprint_all();
+
+        println!("[BOT LOG] {} is ready.", ready.user.name);
+    }
+
+    async fn interaction_create(&self, ctx: Context, interaction: Interaction) {
+        if interaction.kind() == InteractionType::ApplicationCommand {
+            let interaction = interaction.application_command().unwrap();
+            let result = BotCommandManager::run_command(&ctx, &interaction, &DATABASE).await;
+            result.eprint_all();
+        }
+    }
+}
